requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
playwright>=1.40.0
openai>=1.3.0
httpx>=0.25.0
pydantic>=2.5.0
python-dotenv>=1.0.0
pandas>=2.2.0
numpy>=1.26.0
matplotlib>=3.8.0
plotly>=5.18.0
kaleido>=0.2.1
PyPDF2>=3.0.1
pdfplumber>=0.10.3
Pillow>=10.2.0
beautifulsoup4>=4.12.2
lxml>=4.9.3
aiofiles>=23.2.1

